pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "veda"
version = "0.1.0"
description = "Claude Code Orchestrator - Multi-instance coordination with AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp",
    "pyyaml",
    "textual",
    "rich",
]

[project.scripts]
veda = "veda:main_sync"

[tool.setuptools]
packages = ["veda"]
